requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.1.9
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
inflection==0.5.1
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.2
